+++ packages/riverpod/lib/src/builders.dart
@@ -213,6 +213,106 @@ class AutoDisposeNotifierProviderFamilyBuilder {
   }
 }
 
+/// Builds a [StreamNotifierProvider].
+class StreamNotifierProviderBuilder {
+  /// Builds a [StreamNotifierProvider].
+  const StreamNotifierProviderBuilder();
+
+  /// {@macro riverpod.autoDispose}
+  StreamNotifierProvider<NotifierT, T>
+      call<NotifierT extends StreamNotifier<T>, T>(
+    NotifierT Function() create, {
+    String? name,
+    Iterable<ProviderOrFamily>? dependencies,
+  }) {
+    return StreamNotifierProvider<NotifierT, T>(
+      create,
+      name: name,
+      dependencies: dependencies,
+    );
+  }
+
+  /// {@macro riverpod.autoDispose}
+  AutoDisposeStreamNotifierProviderBuilder get autoDispose {
+    return const AutoDisposeStreamNotifierProviderBuilder();
+  }
+
+  /// {@macro riverpod.family}
+  StreamNotifierProviderFamilyBuilder get family {
+    return const StreamNotifierProviderFamilyBuilder();
+  }
+}
+
+/// Builds a [StreamNotifierProviderFamily].
+class StreamNotifierProviderFamilyBuilder {
+  /// Builds a [StreamNotifierProviderFamily].
+  const StreamNotifierProviderFamilyBuilder();
+
+  /// {@macro riverpod.family}
+  StreamNotifierProviderFamily<NotifierT, T, Arg>
+      call<NotifierT extends FamilyStreamNotifier<T, Arg>, T, Arg>(
+    NotifierT Function() create, {
+    String? name,
+    Iterable<ProviderOrFamily>? dependencies,
+  }) {
+    return StreamNotifierProviderFamily<NotifierT, T, Arg>(
+      create,
+      name: name,
+      dependencies: dependencies,
+    );
+  }
+
+  /// {@macro riverpod.autoDispose}
+  AutoDisposeStreamNotifierProviderFamilyBuilder get autoDispose {
+    return const AutoDisposeStreamNotifierProviderFamilyBuilder();
+  }
+}
+
+/// Builds a [AutoDisposeStreamNotifierProvider].
+class AutoDisposeStreamNotifierProviderBuilder {
+  /// Builds a [AutoDisposeStreamNotifierProvider].
+  const AutoDisposeStreamNotifierProviderBuilder();
+
+  /// {@macro riverpod.autoDispose}
+  AutoDisposeStreamNotifierProvider<NotifierT, T>
+      call<NotifierT extends AutoDisposeStreamNotifier<T>, T>(
+    NotifierT Function() create, {
+    String? name,
+    Iterable<ProviderOrFamily>? dependencies,
+  }) {
+    return AutoDisposeStreamNotifierProvider<NotifierT, T>(
+      create,
+      name: name,
+      dependencies: dependencies,
+    );
+  }
+
+  /// {@macro riverpod.family}
+  AutoDisposeStreamNotifierProviderFamilyBuilder get family {
+    return const AutoDisposeStreamNotifierProviderFamilyBuilder();
+  }
+}
+
+/// Builds a [AutoDisposeStreamNotifierProviderFamily].
+class AutoDisposeStreamNotifierProviderFamilyBuilder {
+  /// Builds a [AutoDisposeStreamNotifierProviderFamily].
+  const AutoDisposeStreamNotifierProviderFamilyBuilder();
+
+  /// {@macro riverpod.family}
+  AutoDisposeStreamNotifierProviderFamily<NotifierT, T, Arg>
+      call<NotifierT extends AutoDisposeFamilyStreamNotifier<T, Arg>, T, Arg>(
+    NotifierT Function() create, {
+    String? name,
+    Iterable<ProviderOrFamily>? dependencies,
+  }) {
+    return AutoDisposeStreamNotifierProviderFamily<NotifierT, T, Arg>(
+      create,
+      name: name,
+      dependencies: dependencies,
+    );
+  }
+}
+
 /// Builds a [StateProvider].
 class StateProviderBuilder {
   /// Builds a [StateProvider].