+++ packages/riverpod/CHANGELOG.md
@@ -1,5 +1,21 @@
 ## Unreleased minor
 
+- Added `StreamNotifier` + `StreamNotifierProvider`.
+  This is for building a `StreamProvider` while exposing ways to modify the stream.
+
+  It is primarily meant to be used using code-generation via riverpod_generator,
+  by writing:
+
+  ```dart
+  @riverpod
+  class Example extends _$Example {
+    @override
+    Stream<Model> build() {
+      // TODO return some stream
+    }
+  }
+  ```
+
 - Deprecated `StreamProvider.stream`
   Instead of:
 
@@ -24,7 +40,7 @@
   Do:
 
   ```dart
-  final a = FuturProvider((ref) async {
+  final a = FutureProvider((ref) async {
     final e = await ref.watch(b.future);
     return Model(e);
   })